DBZ-8959 Move events to SPI runtime module
