+++ quarkus-debezium-parent/quarkus-debezium-engine-spi-parent/runtime/src/main/java/io/debezium/runtime/events/AbstractDebeziumLifecycleEvent.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.quarkus.debezium.engine.events;
+package io.debezium.runtime.events;
 
 import io.debezium.runtime.Debezium;
