+++ quarkus-debezium-parent/quarkus-debezium-engine-spi-parent/runtime/src/main/java/io/debezium/runtime/events/ConnectorStoppedEvent.java
@@ -3,17 +3,17 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.quarkus.debezium.engine.events;
+package io.debezium.runtime.events;
 
 import io.debezium.runtime.Debezium;
 
 /**
- * This event is fired when Debezium engine stops polling the connector for changes.
+ * This event is fired when Debezium engine stops a connector.
  *
  * @author Chris Cranford
  */
-public class PollingStoppedEvent extends AbstractDebeziumLifecycleEvent {
-    public PollingStoppedEvent(Debezium engine) {
+public class ConnectorStoppedEvent extends AbstractDebeziumLifecycleEvent {
+    public ConnectorStoppedEvent(Debezium engine) {
         super(engine);
     }
 }
