+++ quarkus-debezium-parent/quarkus-debezium-engine-spi-parent/runtime/src/main/java/io/debezium/runtime/events/PollingStoppedEvent.java
@@ -3,17 +3,17 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.quarkus.debezium.engine.events;
+package io.debezium.runtime.events;
 
 import io.debezium.runtime.Debezium;
 
 /**
- * This event is fired when the connector task is stopped.
+ * This event is fired when Debezium engine stops polling the connector for changes.
  *
  * @author Chris Cranford
  */
-public class TasksStoppedEvent extends AbstractDebeziumLifecycleEvent {
-    public TasksStoppedEvent(Debezium engine) {
+public class PollingStoppedEvent extends AbstractDebeziumLifecycleEvent {
+    public PollingStoppedEvent(Debezium engine) {
         super(engine);
     }
 }
